moved a bunch of the helper functions for manipulating canvases to files in js but they are imported to index.js. Also added some things to settings and moved setting up url input to its own file.
