+++ js/invert.js
@@ -1,0 +1,8 @@
+module.exports = function(data) {
+  for (var i = 0; i < data.length; i += 4) {
+    data[i]     = 255 - data[i];     // red
+    data[i + 1] = 255 - data[i + 1]; // green
+    data[i + 2] = 255 - data[i + 2]; // blue
+    data[i + 3] = 255;
+  }
+};
